# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathPilot API - Career exploration platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  gradeLevel?: string | null;
  academicInterests?: string[];
  strongestSubjects?: string[];
  activities?: string[];
  competitions?: string[];
  projects?: string[];
  goals?: string[];
  careerInterests?: string[];
  onboardingComplete: boolean;
  createdAt: string;
}

export interface ProfileInput {
  name: string;
  email: string;
  gradeLevel?: string;
  academicInterests?: string[];
  strongestSubjects?: string[];
  activities?: string[];
  competitions?: string[];
  projects?: string[];
  goals?: string[];
  careerInterests?: string[];
}

export interface ProfileUpdate {
  name?: string;
  gradeLevel?: string;
  academicInterests?: string[];
  strongestSubjects?: string[];
  activities?: string[];
  careerInterests?: string[];
}

export type SimulationDifficulty = typeof SimulationDifficulty[keyof typeof SimulationDifficulty];


export const SimulationDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Simulation {
  id: number;
  title: string;
  careerCategory: string;
  difficulty: SimulationDifficulty;
  estimatedMinutes: number;
  description: string;
  /** @nullable */
  tagline?: string | null;
  skills: string[];
  stages: number;
  completionCount?: number;
  /** @nullable */
  coverColor?: string | null;
}

export type SimulationSessionStatus = typeof SimulationSessionStatus[keyof typeof SimulationSessionStatus];


export const SimulationSessionStatus = {
  active: 'active',
  completed: 'completed',
  abandoned: 'abandoned',
} as const;

export type SimulationSessionMetrics = {
  analyticalThinking?: number;
  communication?: number;
  problemSolving?: number;
  leadership?: number;
  adaptability?: number;
};

export interface Decision {
  id: number;
  sessionId: number;
  stageNumber: number;
  choiceId: string;
  choiceText: string;
  consequence: string;
  scoreImpact: number;
}

export type ChoiceRisk = typeof ChoiceRisk[keyof typeof ChoiceRisk];


export const ChoiceRisk = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Choice {
  id: string;
  text: string;
  risk: ChoiceRisk;
  /** @nullable */
  hint?: string | null;
}

export interface Scenario {
  stageNumber: number;
  title: string;
  description: string;
  context: string;
  /** @nullable */
  timeLimit?: number | null;
  choices: Choice[];
}

export interface SimulationSession {
  id: number;
  simulationId: number;
  /** @nullable */
  simulationTitle?: string | null;
  status: SimulationSessionStatus;
  currentStage: number;
  totalStages: number;
  score: number;
  decisions: Decision[];
  metrics: SimulationSessionMetrics;
  currentScenario?: Scenario;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface SimulationSessionInput {
  simulationId: number;
}

export interface DecisionInput {
  choiceId: string;
  stageNumber: number;
}

export interface CareerAnalysis {
  id: number;
  sessionId: number;
  primaryCareer: string;
  primaryCareerCategory?: string;
  secondaryCareer: string;
  secondaryCareerCategory?: string;
  confidenceScore: number;
  behavioralInsights: string[];
  strengths: string[];
  growthAreas: string[];
  summary: string;
  decisionPatterns?: string[];
}

export type OpportunityCategory = typeof OpportunityCategory[keyof typeof OpportunityCategory];


export const OpportunityCategory = {
  internship: 'internship',
  research: 'research',
  scholarship: 'scholarship',
  competition: 'competition',
  hackathon: 'hackathon',
  mentorship: 'mentorship',
  summer_program: 'summer_program',
} as const;

export type OpportunityDifficulty = typeof OpportunityDifficulty[keyof typeof OpportunityDifficulty];


export const OpportunityDifficulty = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Opportunity {
  id: number;
  title: string;
  organization: string;
  category: OpportunityCategory;
  relevanceScore: number;
  whyItMatches: string;
  difficulty: OpportunityDifficulty;
  applicationDeadline: string;
  description: string;
  /** @nullable */
  stipend?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  url?: string | null;
  careerCategories?: string[];
}

export type RoadmapMilestoneType = typeof RoadmapMilestoneType[keyof typeof RoadmapMilestoneType];


export const RoadmapMilestoneType = {
  course: 'course',
  skill: 'skill',
  project: 'project',
  certification: 'certification',
  competition: 'competition',
  experience: 'experience',
} as const;

export type RoadmapMilestonePriority = typeof RoadmapMilestonePriority[keyof typeof RoadmapMilestonePriority];


export const RoadmapMilestonePriority = {
  essential: 'essential',
  recommended: 'recommended',
  optional: 'optional',
} as const;

export interface RoadmapMilestone {
  id: number;
  title: string;
  type: RoadmapMilestoneType;
  timeframe: string;
  description: string;
  priority: RoadmapMilestonePriority;
  resources?: string[];
}

export type CareerRoadmapPhasesItem = {
  name: string;
  timeframe: string;
  focus: string;
};

export interface CareerRoadmap {
  targetCareer: string;
  targetCareerCategory?: string;
  milestones: RoadmapMilestone[];
  phases: CareerRoadmapPhasesItem[];
}

export interface Dashboard {
  profileComplete: boolean;
  simulationsCompleted: number;
  /** @nullable */
  topCareerMatch: string | null;
  /** @nullable */
  topCareerConfidence?: number | null;
  activeOpportunities: number;
  roadmapProgress: number;
  recentSessions: SimulationSession[];
  currentStreak?: number;
  totalScore?: number;
}

export type ListOpportunitiesParams = {
/**
 * Filter by category (internship, research, scholarship, competition, hackathon, mentorship, summer_program)
 * @nullable
 */
category?: string | null;
};