# lib/api-zod/src/generated/types/careerAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathPilot API - Career exploration platform
 * OpenAPI spec version: 0.1.0
 */

export interface CareerAnalysis {
  id: number;
  sessionId: number;
  primaryCareer: string;
  primaryCareerCategory?: string;
  secondaryCareer: string;
  secondaryCareerCategory?: string;
  confidenceScore: number;
  behavioralInsights: string[];
  strengths: string[];
  growthAreas: string[];
  summary: string;
  decisionPatterns?: string[];
}