# lib/api-zod/src/generated/types/choice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathPilot API - Career exploration platform
 * OpenAPI spec version: 0.1.0
 */
import type { ChoiceRisk } from './choiceRisk';

export interface Choice {
  id: string;
  text: string;
  risk: ChoiceRisk;
  /** @nullable */
  hint?: string | null;
}