# lib/api-zod/src/generated/types/opportunity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathPilot API - Career exploration platform
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityCategory } from './opportunityCategory';
import type { OpportunityDifficulty } from './opportunityDifficulty';

export interface Opportunity {
  id: number;
  title: string;
  organization: string;
  category: OpportunityCategory;
  relevanceScore: number;
  whyItMatches: string;
  difficulty: OpportunityDifficulty;
  applicationDeadline: string;
  description: string;
  /** @nullable */
  stipend?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  url?: string | null;
  careerCategories?: string[];
}