# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PathPilot API - Career exploration platform
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  gradeLevel?: string | null;
  academicInterests?: string[];
  strongestSubjects?: string[];
  activities?: string[];
  competitions?: string[];
  projects?: string[];
  goals?: string[];
  careerInterests?: string[];
  onboardingComplete: boolean;
  createdAt: string;
}